requirements.txt
pyyaml>=6.0
httpx>=0.25.0
websockets>=12.0
pynacl>=1.5.0
eth-account>=0.11.0
